Fix request tests naming
