+++ client/rest-high-level/src/test/java/org/elasticsearch/client/ml/GetBucketsRequestTests.java
@@ -18,36 +18,47 @@
  */
 package org.elasticsearch.client.ml;
 
-import org.elasticsearch.common.unit.TimeValue;
+import org.elasticsearch.client.ml.job.util.PageParams;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.test.AbstractXContentTestCase;
 
 import java.io.IOException;
 
-public class GetBucketsRequestTests extends AbstractXContentTestCase<GetOverallBucketsRequest> {
+public class GetBucketsRequestTests extends AbstractXContentTestCase<GetBucketsRequest> {
 
     @Override
-    protected GetOverallBucketsRequest createTestInstance() {
-        GetOverallBucketsRequest request = new GetOverallBucketsRequest(randomAlphaOfLengthBetween(1, 20));
+    protected GetBucketsRequest createTestInstance() {
+        GetBucketsRequest request = new GetBucketsRequest(randomAlphaOfLengthBetween(1, 20));
 
         if (randomBoolean()) {
-            request.setTopN(randomIntBetween(1, 10));
+            request.setTimestamp(String.valueOf(randomLong()));
+        } else {
+            if (randomBoolean()) {
+                request.setStart(String.valueOf(randomLong()));
+            }
+            if (randomBoolean()) {
+                request.setEnd(String.valueOf(randomLong()));
+            }
+            if (randomBoolean()) {
+                request.setExcludeInterim(randomBoolean());
+            }
+            if (randomBoolean()) {
+                request.setAnomalyScore(randomDouble());
+            }
+            if (randomBoolean()) {
+                int from = randomInt(10000);
+                int size = randomInt(10000);
+                request.setPageParams(new PageParams(from, size));
+            }
+            if (randomBoolean()) {
+                request.setSort("anomaly_score");
+            }
+            if (randomBoolean()) {
+                request.setDescending(randomBoolean());
+            }
         }
-
         if (randomBoolean()) {
-            request.setBucketSpan(TimeValue.timeValueSeconds(randomIntBetween(1, 1_000_000)));
-        }
-        if (randomBoolean()) {
-            request.setStart(String.valueOf(randomLong()));
-        }
-        if (randomBoolean()) {
-            request.setEnd(String.valueOf(randomLong()));
-        }
-        if (randomBoolean()) {
-            request.setExcludeInterim(randomBoolean());
-        }
-        if (randomBoolean()) {
-            request.setOverallScore(randomDouble());
+            request.setExpand(randomBoolean());
         }
         if (randomBoolean()) {
             request.setExcludeInterim(randomBoolean());
@@ -56,8 +67,8 @@ public class GetBucketsRequestTests extends AbstractXContentTestCase<GetOverallB
     }
 
     @Override
-    protected GetOverallBucketsRequest doParseInstance(XContentParser parser) throws IOException {
-        return GetOverallBucketsRequest.PARSER.apply(parser, null);
+    protected GetBucketsRequest doParseInstance(XContentParser parser) throws IOException {
+        return GetBucketsRequest.PARSER.apply(parser, null);
     }
 
     @Override
